refactor: move LogWrapper() to common library
Just moving some helper functions from `spanner/internal` to
`google/cloud/internal`. These functions are used to implement logging
decorators for the stubs.
