+++ google/cloud/spanner/internal/instance_admin_logging.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/spanner/internal/instance_admin_logging.h"
-#include "google/cloud/spanner/internal/log_wrapper.h"
+#include "google/cloud/internal/log_wrapper.h"
 
 namespace google {
 namespace cloud {
@@ -22,6 +22,7 @@ inline namespace SPANNER_CLIENT_NS {
 namespace internal {
 
 namespace gcsa = ::google::spanner::admin::instance::v1;
+using ::google::cloud::internal::LogWrapper;
 
 StatusOr<gcsa::Instance> InstanceAdminLogging::GetInstance(
     grpc::ClientContext& context, gcsa::GetInstanceRequest const& request) {
