+++ google/cloud/spanner/spanner_client.bzl
@@ -46,7 +46,6 @@ spanner_client_hdrs = [
     "internal/instance_admin_logging.h",
     "internal/instance_admin_metadata.h",
     "internal/instance_admin_stub.h",
-    "internal/log_wrapper.h",
     "internal/logging_result_set_reader.h",
     "internal/logging_spanner_stub.h",
     "internal/merge_chunk.h",
@@ -106,7 +105,6 @@ spanner_client_srcs = [
     "internal/instance_admin_logging.cc",
     "internal/instance_admin_metadata.cc",
     "internal/instance_admin_stub.cc",
-    "internal/log_wrapper.cc",
     "internal/logging_result_set_reader.cc",
     "internal/logging_spanner_stub.cc",
     "internal/merge_chunk.cc",
